pyproject.toml
[tool.poetry]
name = "py-project"
version = "0.1.0"
description = ""
authors = ["David Mugisha"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.4"
pandera = "^0.12.0"
fsspec = "^2022.8.2"
SQLAlchemy = "^1.4.40"
adlfs = "^2022.9.1"
psycopg2-binary = "^2.9.3"
azure-functions-durable = "^1.1.6"
azure-functions = "^1.11.2"
psutil = "^5.9.3"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
flake8-formatter-junit-xml = "^0.0.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
bandit = "^1.7.4"
pre-commit = "^2.20.0"
isort = "^5.10.1"
safety = "^2.3.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
"testing.postgresql" = "^1.3.0"

[tool.pytest.ini_options]
junit_family = "xunit1"

[tool.coverage.run]
source=["py_project/*"]
omit=["tests/*","venv/*"]

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"

[tool.bandit]
exclude_dirs = ["tests", "docker", ".venv"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
